should prob iterate with `size_t`?

+++ src/main.c
@@ -32,17 +32,17 @@ int main(void) {
     tpool_t pool;
     tpool_set(&pool, N_THREADS);
     uint16_t data[N_ITEMS] = {0};
-    for (uint16_t i = 0; i < N_ITEMS; ++i) {
-        data[i] = i;
+    for (size_t i = 0; i < N_ITEMS; ++i) {
+        data[i] = (uint16_t)i;
     }
-    for (uint16_t i = 0; i < N_ITEMS; ++i) {
+    for (size_t i = 0; i < N_ITEMS; ++i) {
         EXIT_IF(!tpool_work_enqueue(&pool, worker, &data[i]))
     }
     usleep(10); /* NOTE: The pool sometimes doesn't kick off. This tiny pause
                  * seems to help... but, why? What is going on here?
                  */
     tpool_wait(&pool);
-    for (uint16_t i = 0; i < N_ITEMS; ++i) {
+    for (size_t i = 0; i < N_ITEMS; ++i) {
         printf("%hu\n", data[i]);
     }
     tpool_clear(&pool);
